requirements.txt
# Retail Sales Agent System Requirements

# Google Generative AI and ADK
google-generativeai>=0.8.0
google-adk>=0.1.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlite3

# Utilities
python-dotenv>=1.0.0

# Data handling
typing-extensions>=4.8.0

# Optional: for future enhancements
flask>=3.0.0  # For REST API
streamlit>=1.28.0  # For UI demo
